req1_llm_only/requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
